Moving functions to context
