+++ src/hooks/index.ts
@@ -26,6 +26,7 @@ import usePrevious from './usePrevious';
 import useWalletState from './useWalletState';
 import useMinkeRewards from './useMinkeRewards';
 import useWalletManagement from './useWalletManagement';
+import useNFT from './useNFT';
 
 export {
 	useCountry,
@@ -55,5 +56,6 @@ export {
 	useFirstTime,
 	usePrevious,
 	useMinkeRewards,
-	useWalletManagement
+	useWalletManagement,
+	useNFT
 };
