+++ src/hooks/useNFT.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { NFTContext } from '../contexts/NFTContext/NFTContext';
+
+const useNFT = () => useContext(NFTContext);
+
+export default useNFT;
